server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-cloud-vision==3.5.0
python-dotenv==1.0.1